requirements.txt
Mako==1.0.3
MarkupSafe==0.23
pycparser==2.14
Pygments==2.1
pyramid==1.8.1
pyramid-beaker==0.8
pyramid-mako==1.0.2
Unidecode==0.4.19
waitress==0.8.10
bcrypt==3.1.3
python-slugify==1.2.1